feat: Basic Statements exercises part 3 completed

+++ albums_migration.sql
@@ -16,7 +16,7 @@ CREATE TABLE albums (
     artist VARCHAR(100),
     name VARCHAR(100) NOT NULL,
     release_date SMALLINT UNSIGNED NOT NULL,
-    sales FLOAT(2) UNSIGNED,
+    sales FLOAT UNSIGNED,
     genre VARCHAR(100),
     PRIMARY KEY (id)
 );
